+++ src_new/triggers/levi_ataxia/for_levi/leviticus/highlighting/061_Corpse_Eaten.lua
@@ -32,7 +32,7 @@ colorTriggerBgColor: '#000000'
 patterns:
 - pattern: you dig into the corpse of
   type: 0
-- pattern: Your fatigue fades away as the meat slides down your gullet\.
+- pattern: Your fatigue fades away as the meat slides down your gullet.
   type: 0
 ]]--